Create README.md
This project mainly focuses on handling large dataset and process the data with an efficient manner. The technology
used in this project was Hadoop MapReduce to make fault-tolerant storage (hdfs) with high throughput. With this
paper, the research made on handling massive amount of data with MapReduce framework and big data. Applied
these techniques on large scale data from information and communication sector (ICT) and passenger movement to
see if there is a correlation occurs between the economic growth of ICT sector and Ireland’s travel report.